requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
mysqlclient
gunicorn